+++ src/app/containers/Preference/Snippet/collectionInfoModal/style.tsx
@@ -1,1 +1,3 @@
 export const labelStyle = { width: '80%', marginBottom: 15 };
+
+export const btnStyle = { fontSize: 14 };